Add files via upload
The Library class was created as requested, along with a Main class that interacts with it and two supporting classes - Action and ActionDetails - which are used for the serializing/deserializing of the JSON strings and contain information for the user-entered actions, respectively. Finally, a LibraryTest class was created and includes unit tests for the Library class.